refactor so we can use for edit
