+++ client/src/app/Authorized/PageContent/Transactions/TransactionsHeader/TransactionsSettings/AutomaticRules/AddAutomaticRule/AddAutomaticRule.tsx
@@ -1,28 +1,26 @@
-import { Button, Stack, Group, Text, ActionIcon } from "@mantine/core";
+import { Button, Stack } from "@mantine/core";
 import { notifications } from "@mantine/notifications";
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { AxiosError } from "axios";
 import React from "react";
-import { AuthContext } from "~/components/AuthProvider/AuthProvider";
+
 import { translateAxiosError } from "~/helpers/requests";
 import {
   FieldToOperatorType,
   IAutomaticCategorizationRuleRequest,
-  IRuleParameterCreateRequest,
+  IRuleParameterEdit,
   Operators,
   TransactionFields,
 } from "~/models/automaticCategorizationRule";
-import { ICategoryResponse } from "~/models/category";
-import { defaultTransactionCategories } from "~/models/transaction";
-import ConditionItem from "./ConditionItem/ConditionItem";
-import ActionItem from "./ActionItem/ActionItem";
-import { PlusIcon } from "lucide-react";
+
+import EditableAutomaticRuleContent from "../EditableAutomaticRuleContent/EditableAutomaticRuleContent";
+import { AuthContext } from "~/components/AuthProvider/AuthProvider";
 
 const AddAutomaticRule = (): React.ReactNode => {
   const defaultField =
     TransactionFields.find((field) => field.value === "merchant")?.value ?? "";
   const [conditionItems, setConditionItems] = React.useState<
-    IRuleParameterCreateRequest[]
+    IRuleParameterEdit[]
   >([
     {
       field: defaultField,
@@ -33,13 +31,10 @@ const AddAutomaticRule = (): React.ReactNode => {
           .map((op) => op.value)
           .at(0) ?? "",
       value: "",
-      type: "",
     },
   ]);
 
-  const [actionItems, setActionItems] = React.useState<
-    IRuleParameterCreateRequest[]
-  >([
+  const [actionItems, setActionItems] = React.useState<IRuleParameterEdit[]>([
     {
       field: defaultField,
       operator:
@@ -47,31 +42,10 @@ const AddAutomaticRule = (): React.ReactNode => {
           (op) => op.type === FieldToOperatorType.get(defaultField)
         )?.value ?? "",
       value: "",
-      type: "",
     },
   ]);
 
   const { request } = React.useContext<any>(AuthContext);
-
-  const transactionCategoriesQuery = useQuery({
-    queryKey: ["transactionCategories"],
-    queryFn: async () => {
-      const res = await request({
-        url: "/api/transactionCategory",
-        method: "GET",
-      });
-
-      if (res.status === 200) {
-        return res.data as ICategoryResponse[];
-      }
-
-      return undefined;
-    },
-  });
-
-  const transactionCategoriesWithCustom = defaultTransactionCategories.concat(
-    transactionCategoriesQuery.data ?? []
-  );
 
   const queryClient = useQueryClient();
   const doAddRule = useMutation({
@@ -97,112 +71,30 @@ const AddAutomaticRule = (): React.ReactNode => {
     },
   });
 
-  const addNewCondition = () => {
-    setConditionItems((prev) => [
-      ...prev,
-      {
-        field: defaultField,
-        operator:
-          Operators.filter(
-            (op) => op.type === FieldToOperatorType.get(defaultField)
-          )
-            .map((op) => op.value)
-            .at(0) ?? "",
-        value: "",
-        type: "",
-      },
-    ]);
-  };
-
-  const removeCondition = (index: number) => {
-    setConditionItems((prev) => prev.filter((_, i) => i !== index));
-  };
-
-  const addNewAction = () => {
-    setActionItems((prev) => [
-      ...prev,
-      {
-        field: TransactionFields.at(0)?.value ?? "",
-        operator:
-          Operators.find(
-            (op) => op.type === FieldToOperatorType.get(defaultField)
-          )?.value ?? "",
-        value: "",
-        type: "",
-      },
-    ]);
-  };
-
-  const removeAction = (index: number) => {
-    setActionItems((prev) => prev.filter((_, i) => i !== index));
-  };
-
   return (
     <Stack gap="0.5rem">
-      <Stack gap="0.5rem">
-        <Group align="center" justify="space-between">
-          <Text size="sm" fw={600}>
-            If
-          </Text>
-          <ActionIcon size="sm" onClick={addNewCondition}>
-            <PlusIcon size={16} />
-          </ActionIcon>
-        </Group>
-        {conditionItems.map(
-          (item: IRuleParameterCreateRequest, index: number) => (
-            <ConditionItem
-              key={index}
-              ruleParameter={item}
-              setRuleParameter={(newParameter) =>
-                setConditionItems(
-                  (
-                    prev: IRuleParameterCreateRequest[]
-                  ): IRuleParameterCreateRequest[] =>
-                    prev.map((param, i) => (i === index ? newParameter : param))
-                )
-              }
-              allowDelete={conditionItems.length > 1}
-              doDelete={removeCondition}
-              index={index}
-              categories={transactionCategoriesWithCustom}
-            />
-          )
-        )}
-      </Stack>
-      <Stack gap="0.5rem">
-        <Group align="center" justify="space-between">
-          <Text size="sm" fw={600}>
-            Then
-          </Text>
-          <ActionIcon size="sm" onClick={addNewAction}>
-            <PlusIcon size={16} />
-          </ActionIcon>
-        </Group>
-        {actionItems.map((item: IRuleParameterCreateRequest, index: number) => (
-          <ActionItem
-            key={index}
-            ruleParameter={item}
-            setRuleParameter={(newParameter) =>
-              setActionItems(
-                (
-                  prev: IRuleParameterCreateRequest[]
-                ): IRuleParameterCreateRequest[] =>
-                  prev.map((param, i) => (i === index ? newParameter : param))
-              )
-            }
-            allowDelete={actionItems.length > 1}
-            doDelete={removeAction}
-            index={index}
-            categories={transactionCategoriesWithCustom}
-          />
-        ))}
-      </Stack>
+      <EditableAutomaticRuleContent
+        conditionItems={conditionItems}
+        actionItems={actionItems}
+        setConditionItems={setConditionItems}
+        setActionItems={setActionItems}
+      />
       <Button
         loading={doAddRule.isPending}
         onClick={() => {
           doAddRule.mutate({
-            conditions: conditionItems,
-            actions: actionItems,
+            conditions: conditionItems.map((item) => ({
+              field: item.field,
+              operator: item.operator,
+              value: item.value,
+              type: "",
+            })),
+            actions: actionItems.map((item) => ({
+              field: item.field,
+              operator: item.operator,
+              value: item.value,
+              type: "",
+            })),
           });
           setConditionItems([
             {
@@ -214,7 +106,6 @@ const AddAutomaticRule = (): React.ReactNode => {
                   .map((op) => op.value)
                   .at(0) ?? "",
               value: "",
-              type: "",
             },
           ]);
           setActionItems([
@@ -225,7 +116,6 @@ const AddAutomaticRule = (): React.ReactNode => {
                   (op) => op.type === FieldToOperatorType.get(defaultField)
                 )?.value ?? "",
               value: "",
-              type: "",
             },
           ]);
         }}
