+++ client/src/app/Authorized/PageContent/Transactions/TransactionsHeader/TransactionsSettings/AutomaticRules/EditableAutomaticRuleContent/ConditionItem/ConditionItem.tsx
@@ -4,7 +4,6 @@ import {
   Group,
   NumberInput,
   Select,
-  Text,
   TextInput,
 } from "@mantine/core";
 import { DateInput } from "@mantine/dates";
@@ -16,22 +15,24 @@ import { AuthContext } from "~/components/AuthProvider/AuthProvider";
 import CategorySelect from "~/components/CategorySelect";
 import { getCurrencySymbol } from "~/helpers/currency";
 import {
-  IRuleParameterCreateRequest,
+  FieldToOperatorType,
+  IRuleParameterEdit,
+  Operators,
   TransactionFields,
 } from "~/models/automaticCategorizationRule";
 import { ICategory } from "~/models/category";
 import { IUserSettings } from "~/models/userSettings";
 
-export interface ActionItemProps {
-  ruleParameter: IRuleParameterCreateRequest;
-  setRuleParameter: (newParameter: IRuleParameterCreateRequest) => void;
-  allowDelete: boolean;
-  doDelete: (index: number) => void;
+export interface ConditionItemProps {
+  ruleParameter: IRuleParameterEdit;
+  setRuleParameter: (newParameter: IRuleParameterEdit) => void;
+  allowDelete?: boolean;
+  doDelete?: (index: number) => void;
   index: number;
   categories: ICategory[];
 }
 
-const ActionItem = (props: ActionItemProps): React.ReactNode => {
+const ConditionItem = (props: ConditionItemProps): React.ReactNode => {
   const { request } = React.useContext<any>(AuthContext);
 
   const userSettingsQuery = useQuery({
@@ -119,9 +120,6 @@ const ActionItem = (props: ActionItemProps): React.ReactNode => {
   return (
     <Card p="0.5rem" radius="md">
       <Group gap="0.5rem">
-        <Text size="sm" fw={600}>
-          Set
-        </Text>
         <Select
           w="110px"
           data={TransactionFields.map((field) => field.label)}
@@ -136,20 +134,46 @@ const ActionItem = (props: ActionItemProps): React.ReactNode => {
               field:
                 TransactionFields.find((field) => field.label === value)
                   ?.value ?? "",
+              operator:
+                Operators.filter(
+                  (op) =>
+                    op.type ===
+                    FieldToOperatorType.get(
+                      TransactionFields.find((field) => field.label === value)
+                        ?.value ?? ""
+                    )
+                ).at(0)?.value ?? "",
+              value: "",
             })
           }
+          allowDeselect={false}
         />
-        <Text size="sm" fw={600}>
-          to
-        </Text>
+        <Select
+          w="130px"
+          data={Operators.filter(
+            (op) =>
+              op.type === FieldToOperatorType.get(props.ruleParameter.field)
+          ).map((op) => op.label)}
+          value={
+            Operators.find((op) => op.value === props.ruleParameter.operator)
+              ?.label ?? ""
+          }
+          onChange={(value) =>
+            props.setRuleParameter({
+              ...props.ruleParameter,
+              operator: Operators.find((op) => op.label === value)?.value ?? "",
+            })
+          }
+          allowDeselect={false}
+        />
         {getValueInput()}
         {props.allowDelete && (
           <Group style={{ alignSelf: "stretch" }}>
             <ActionIcon
-              h="100%"
-              size="sm"
               color="red"
-              onClick={() => props.doDelete(props.index)}
+              size="sm"
+              h="100%"
+              onClick={() => props.doDelete?.(props.index)}
             >
               <Trash2Icon size={16} />
             </ActionIcon>
@@ -160,4 +184,4 @@ const ActionItem = (props: ActionItemProps): React.ReactNode => {
   );
 };
 
-export default ActionItem;
+export default ConditionItem;
